# java/src/main/java/com/silver/apichallenge/entity/User.java
package com.silver.apichallenge.entity;

import com.fasterxml.jackson.annotation.JsonProperty;

import java.util.List;

public class User {
    private String id;
    @JsonProperty("nome")
    private String name;
    @JsonProperty("idade")
    private Integer age;
    private Integer score;
    @JsonProperty("ativo")
    private Boolean active;
    @JsonProperty("pais")
    private String country;
    @JsonProperty("equipe")
    private Team team;
    private List<Log> logs;

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Integer getAge() {
        return age;
    }

    public void setAge(Integer age) {
        this.age = age;
    }

    public Integer getScore() {
        return score;
    }

    public void setScore(Integer score) {
        this.score = score;
    }

    public Boolean getActive() {
        return active;
    }

    public void setActive(Boolean active) {
        this.active = active;
    }

    public String getCountry() {
        return country;
    }

    public void setCountry(String country) {
        this.country = country;
    }

    public Team getTeam() {
        return team;
    }

    public void setTeam(Team team) {
        this.team = team;
    }

    public List<Log> getLogs() {
        return logs;
    }

    public void setLogs(List<Log> logs) {
        this.logs = logs;
    }
}